work/requirements.txt
psycopg2-binary
redis
bokeh
plotly
chart_studio
numpy
scipy
python-dotenv
pandas
spotipy
dash
pyspark